Multiplication Table Generator USING ALL 3 LOOPS
his repository contains a C++ program that generates multiplication tables for a given number using three different types of loops: for-loop, while-loop, and do-while-loop. The user is prompted to enter the number for which they want the multiplication table, and the program displays the table from 1 to 10 using each of the three loops. This program can be useful for anyone who wants to practice or teach multiplication tables in a fun and interactive way.